Store P* instead of harmony, renamed stuff
Rather than storing harmonies in dictionary, it makes more sense to
store maxent values (P*), as these are used both in calculating Z scores
and in calculating Probabilities.

Also, renamed functions to be in line with Python standards (lowercase
with underscores instead of CamelCase), and renamed some variables for
the sake of transparency.