Test module to quickly see if code works
You can run tests.py to see if the code is currently working. If it is,
it'll return two probabilities and won't throw any errors.

It currently throws errors when toy_input_2 is used, because this file
doesn't mark 0 violations, and megatableau currently doesn't add 0s in
during read-in like it used to do.