Fix the L1/L2 priors
I took the example formulas that Robert had added in in-line comments and used them for calculating the actual gradients, fiddling with signs until they seemed intuitively reasonable and the learned weights looked good.  I also made the prior variable names more consistent, and set the bounds to (-50, 0) instead of (-25, 0), as suggested in class.