learn_weights now uses consistent starting weights
In the learn_weights() function, initial weights are set to zero rather
than being randomized.

Doing this makes the results of maxent.py and phlearn.py consistent from
call to call (for better or worse).

The wild fluctuations we'd been observing for some tableau / training
sets (but not others) apparently were a result of the random starting
locations. Not sure what to make of this.