Fixed round-off problem in maxent_value()
Occasionally maxent values of 0 were being created due to round-off
error. This lead to math domain errors in nlpwg(), because of its call
to math.exp(probability), which isn't defined when the probability is 0.

Fixed by adding a very very tiny number to every maxent value, to ensure
positive polarity.